+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to the [Haskell Package Versioning Policy](https://pvp.haskell.org/).
 
 ## [Unreleased]
+- Make `sort` param take a list of asc/desc-tagged field names
 
 ## [0.4.0.1] - 2017-09-20
 ### Changed